Count current tick using an int instead of LongAdder
Assuming 20tps, there's enough space in an int to store a tick count spanning a number of years :)

+++ spark-common/src/main/java/me/lucko/spark/sampler/TickCounter.java
@@ -41,7 +41,7 @@ public interface TickCounter extends AutoCloseable {
      *
      * @return the current tick
      */
-    long getCurrentTick();
+    int getCurrentTick();
 
     /**
      * Adds a task to be called each time the tick increments
